feat: updated cql filters, code cleanup
